# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/business-owners": {
      "get": {
        "operationId": "BusinessOwnerController_getBusinessOwners",
        "parameters": [
          {
            "name": "companyName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessOwnerGetResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/business-owners/{id}": {
      "get": {
        "operationId": "BusinessOwnerController_getBusinessOwnerByIdForInvestor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOwnerGetDetailedResponse"
                }
              }
            }
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/business-owners/my/profile/image": {
      "put": {
        "operationId": "BusinessOwneProfileController_uploadProfileImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "BusinessOwneProfileController_deleteFileById",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/business-owners/my/profile": {
      "patch": {
        "operationId": "BusinessOwneProfileController_patchProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessOwnerProfilePatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "BusinessOwneProfileController_getBusinessOwnersProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOwnerProfileGetResponse"
                }
              }
            }
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "BusinessOwneProfileController_deleteBusinessOwner",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/business-owners/my/profile/preview": {
      "get": {
        "operationId": "BusinessOwneProfileController_getBusinessOwnerProfilePreview",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOwnerProfilePreviewGetResponse"
                }
              }
            }
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/business-owners/my/profile/monthly-report": {
      "get": {
        "operationId": "BusinessOwneProfileController_getBusinessOwnersMonthlyReport",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyReportRatingResponse"
                }
              }
            }
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "BusinessOwneProfileController_updateBusinessOwnersMonthlyReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMonthlyReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "BusinessOwneProfileController_confirmMonthlyReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmMonthlyReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/business-owners/my/profile/alerts": {
      "get": {
        "operationId": "BusinessOwneProfileController_getBusinessOwnerAlerts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOwnerAlertsResponse"
                }
              }
            }
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/business-owners/my/profile/monthly-report-chart": {
      "get": {
        "operationId": "BusinessOwneProfileController_getMonthlyReportChartData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChartPoint"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "business owners"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin-data/business-owners/{id}": {
      "get": {
        "operationId": "BusinessOwnerAdminDataController_getBusinessOwnerByIdForAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOwnerGetDetailedAdminResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "BusinessOwnerAdminDataController_blockOrUnblockBusinessOwner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockBusinessOwnerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/admin-data/business-owners/{id}/monthly-report": {
      "get": {
        "operationId": "BusinessOwnerAdminDataController_getBusinessOwnersMonthlyReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "BusinessOwnerAdminDataController_decideMonthlyReportUpdate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideMonthlyReport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Asabik API Docs",
    "description": "All endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Address": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "readOnly": true
          },
          "street": {
            "type": "string",
            "readOnly": true
          },
          "zipCode": {
            "type": "string",
            "readOnly": true
          }
        },
        "required": [
          "city",
          "street",
          "zipCode"
        ]
      },
...
      "BusinessGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "BusinessSectorDetailsGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "business": {
            "$ref": "#/components/schemas/BusinessGetResponse"
          }
        },
        "required": [
          "id",
          "name",
          "business"
        ]
      },
      "BusinessSubsectorDetailsGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "businessSector": {
            "$ref": "#/components/schemas/BusinessSectorDetailsGetResponse"
          }
        },
        "required": [
          "id",
          "name",
          "businessSector"
        ]
      },
      "BusinessOwnerGetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "readOnly": true
          },
          "investmentRequestId": {
            "type": "number",
            "readOnly": true
          },
          "isInvested": {
            "type": "boolean",
            "readOnly": true
          },
          "isRepaying": {
            "type": "boolean",
            "readOnly": true
          },
          "investedAmount": {
            "type": "number",
            "readOnly": true
          },
          "companyName": {
            "type": "string",
            "readOnly": true
          },
          "rating": {
            "type": "string",
            "readOnly": true,
            "enum": [
              "Aaa",
              "Aa1",
              "Aa2",
              "Aa3",
              "A1",
              "A2",
              "A3",
              "Baa1",
              "Baa2",
              "Baa3",
              "Junk"
            ]
          },
          "projectedNetReturn": {
            "type": "string",
            "readOnly": true
          },
          "psr": {
            "type": "string",
            "readOnly": true
          },
          "expirationDate": {
            "type": "string",
            "readOnly": true
          },
          "returnTerm": {
            "type": "number",
            "readOnly": true
          },
          "amountToMeetTarget": {
            "type": "number",
            "readOnly": true
          },
          "businessSubsector": {
            "readOnly": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessSubsectorDetailsGetResponse"
              }
            ]
          },
          "businessStructure": {
            "type": "string",
            "readOnly": true
          }
        },
        "required": [
          "id",
          "investmentRequestId",
          "isInvested",
          "isRepaying",
          "investedAmount",
          "companyName",
          "rating",
          "projectedNetReturn",
          "psr",
          "expirationDate",
          "returnTerm",
          "amountToMeetTarget",
          "businessSubsector",
          "businessStructure"
        ]
      },
      "AdressResponse": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "readOnly": true
          },
          "city": {
            "type": "string",
            "readOnly": true
          },
          "zipcode": {
            "type": "string",
            "readOnly": true
          }
        },
        "required": [
          "street",
          "city",
          "zipcode"
        ]
      },
...
      "BusinessOwnerGetDetailedResponse": {
        "type": "object",
        "properties": {
          "imgUrl": {
            "type": "string",
            "readOnly": true
          },
          "companyName": {
            "type": "string",
            "readOnly": true
          },
          "description": {
            "type": "string",
            "readOnly": true
          },
          "ownerName": {
            "type": "string",
            "readOnly": true
          },
          "businessSubsector": {
            "readOnly": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessSubsectorDetailsGetResponse"
              }
            ]
          },
          "startDate": {
            "type": "string",
            "readOnly": true
          },
          "avrMonthlySales": {
            "type": "number",
            "readOnly": true
          },
          "avrMonthlyNetProfit": {
            "type": "number",
            "readOnly": true
          },
          "totalLastYearNetProfit": {
            "type": "number",
            "readOnly": true
          },
          "employeesNo": {
            "type": "number",
            "readOnly": true
          },
          "phone": {
            "type": "string",
            "readOnly": true
          },
          "website": {
            "type": "string",
            "readOnly": true
          },
          "adress": {
            "readOnly": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AdressResponse"
              }
            ]
          },
          "investmentRequests": {
            "readOnly": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvestmentRequestGetInvestorResponse"
            }
          }
        },
        "required": [
          "imgUrl",
          "companyName",
          "description",
          "ownerName",
          "businessSubsector",
          "startDate",
          "avrMonthlySales",
          "avrMonthlyNetProfit",
          "totalLastYearNetProfit",
          "employeesNo",
          "phone",
          "website",
          "adress",
          "investmentRequests"
        ]
      },
      "BusinessOwnerProfilePatchDto": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "BusinessOwnerProfileGetResponse": {
        "type": "object",
        "properties": {
          "imgUrl": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "businessSubsector": {
            "$ref": "#/components/schemas/BusinessSubsectorDetailsGetResponse"
          },
          "businessDescription": {
            "type": "string"
          },
          "investmentRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvestmentRequestGetBusinessOwnerResponse"
            }
          },
          "isProfileComplete": {
            "type": "boolean"
          },
          "isValidForRequest": {
            "type": "boolean"
          }
        },
        "required": [
          "businessSubsector",
          "investmentRequests",
          "isProfileComplete",
          "isValidForRequest"
        ]
      },
      "InvestmentRequestsSummaryResponse": {
        "type": "object",
        "properties": {
          "projectedNetRunUpTo": {
            "type": "number",
            "readOnly": true
          },
          "totalInvestedAmount": {
            "type": "number",
            "readOnly": true
          },
          "numberOfInvestors": {
            "type": "number",
            "readOnly": true
          },
          "totalRequiredInvestment": {
            "type": "number",
            "readOnly": true
          },
          "minReturnTerm": {
            "type": "number",
            "readOnly": true
          }
        },
        "required": [
          "projectedNetRunUpTo",
          "totalInvestedAmount",
          "numberOfInvestors",
          "totalRequiredInvestment",
          "minReturnTerm"
        ]
      },
      "BusinessOwnerProfilePreviewGetResponse": {
        "type": "object",
        "properties": {
          "imgUrl": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "businessSubsector": {
            "$ref": "#/components/schemas/BusinessSubsectorDetailsGetResponse"
          },
          "startDate": {
            "type": "string"
          },
          "avrMonthlySales": {
            "type": "number"
          },
          "avrMonthlyNetProfit": {
            "type": "number"
          },
          "totalLastYearNetProfit": {
            "type": "number"
          },
          "employeesNo": {
            "type": "number"
          },
          "businessDescription": {
            "type": "string"
          },
          "investmentRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvestmentRequestGetInvestorResponse"
            }
          },
          "isProfileComplete": {
            "type": "boolean"
          },
          "isValidForRequest": {
            "type": "boolean"
          },
          "investmentsRequestsSummary": {
            "$ref": "#/components/schemas/InvestmentRequestsSummaryResponse"
          }
        },
        "required": [
          "businessSubsector",
          "startDate",
          "avrMonthlySales",
          "avrMonthlyNetProfit",
          "totalLastYearNetProfit",
          "employeesNo",
          "investmentRequests",
          "isProfileComplete",
          "isValidForRequest",
          "investmentsRequestsSummary"
        ]
      },
      "MonthlyReportRatingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "givenMonth": {
            "type": "string"
          },
          "monthlyIn": {
            "type": "number"
          },
          "monthlyOut": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "givenMonth",
          "monthlyIn",
          "monthlyOut"
        ]
      },
      "BusinessOwnerAlertsResponse": {
        "type": "object",
        "properties": {
          "isReportPending": {
            "type": "boolean",
            "readOnly": true
          },
          "isProfileIncomplete": {
            "type": "boolean",
            "readOnly": true
          },
          "isInvestmentRequestExpired": {
            "type": "boolean",
            "readOnly": true
          },
          "isVerified": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "required": [
          "isReportPending",
          "isProfileIncomplete",
          "isInvestmentRequestExpired",
          "isVerified"
        ]
      },
      "ChartPoint": {
        "type": "object",
        "properties": {
          "reportDateAsX": {
            "type": "string",
            "readOnly": true
          },
          "inflowAsY": {
            "type": "number",
            "readOnly": true
          },
          "outflowAsY": {
            "type": "number",
            "readOnly": true
          }
        },
        "required": [
          "reportDateAsX",
          "inflowAsY",
          "outflowAsY"
        ]
      },
      "UpdateMonthlyReportDto": {
        "type": "object",
        "properties": {
          "monthlyReportId": {
            "type": "number"
          },
          "inflow": {
            "type": "number"
          },
          "inflowDescription": {
            "type": "string"
          },
          "outflow": {
            "type": "number"
          },
          "outflowDescription": {
            "type": "string"
          }
        },
        "required": [
          "inflow",
          "outflow"
        ]
      },
      "ConfirmMonthlyReportDto": {
        "type": "object",
        "properties": {
          "monthlyReportId": {
            "type": "number"
          }
        },
        "required": [
          "monthlyReportId"
        ]
      },
...
      "BusinessOwnerGetDetailedAdminResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "readOnly": true
          },
          "companyName": {
            "type": "string",
            "readOnly": true
          },
          "registrationDate": {
            "format": "date-time",
            "type": "string",
            "readOnly": true
          },
          "ownerName": {
            "type": "string",
            "readOnly": true
          },
          "sector": {
            "type": "string",
            "readOnly": true
          },
          "address": {
            "readOnly": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Address"
              }
            ]
          },
          "receivedInvestments": {
            "type": "number",
            "readOnly": true
          },
          "requiredInvestments": {
            "type": "number",
            "readOnly": true
          },
          "phoneNumber": {
            "type": "string",
            "readOnly": true
          },
          "email": {
            "type": "string",
            "readOnly": true
          },
          "isActive": {
            "type": "boolean",
            "readOnly": true
          },
          "isBlocked": {
            "type": "boolean",
            "readOnly": true
          },
          "isReportAvailable": {
            "type": "boolean",
            "readOnly": true
          },
          "investors": {
            "readOnly": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "investmentRequests": {
            "readOnly": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isVerified": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "required": [
          "id",
          "companyName",
          "registrationDate",
          "ownerName",
          "sector",
          "address",
          "receivedInvestments",
          "requiredInvestments",
          "phoneNumber",
          "email",
          "isActive",
          "isBlocked",
          "isReportAvailable",
          "investors",
          "investmentRequests",
          "isVerified"
        ]
      },
...
      "BlockBusinessOwnerRequest": {
        "type": "object",
        "properties": {
          "isBlocked": {
            "type": "boolean"
          }
        },
        "required": [
          "isBlocked"
        ]
      },
      "DecideMonthlyReport": {
        "type": "object",
        "properties": {
          "monthlyReportId": {
            "type": "number"
          },
          "isApproved": {
            "type": "boolean"
          }
        },
        "required": [
          "monthlyReportId",
          "isApproved"
        ]
      },
...
    }
  }
}